dunyang/requirements.txt
xlsx2html==0.6.4
beautifulsoup4==4.12.3
lxml==5.2.2
pypdfium2==4.30.0
Pillow==10.4.0
